support multiline lookups
This introduces a parameter which allows the plugin to return the full
contents of the pass entry. I use this so that I can store an SSL key in
pass and have ansible deploy it when provisioning a web server. I still
believe that returning only the first line is the correct default.

The "multiline" language should be familiar to pass users who `pass insert
--multiline`.